+++ woodwork/tests/test_utils.py
@@ -164,11 +164,30 @@ def test_list_semantic_tags():
                 assert name in log_type.standard_tags
 
 
-def test_read_csv_no_params(sample_df_pandas, tmpdir):
-    filepath = os.path.join(tmpdir, 'sample.csv')
+def test_read_file_errors_no_content_type(sample_df_pandas, tmpdir):
+    filepath = os.path.join(tmpdir, 'sample')
     sample_df_pandas.to_csv(filepath, index=False)
 
-    df_from_csv = ww.read_csv(filepath=filepath)
+    no_type_error = "Content type could not be inferred. Please specify content_type and try again."
+    with pytest.raises(RuntimeError, match=no_type_error):
+        ww.read_file(filepath=filepath)
+
+
+def test_read_file_errors_unsupported(sample_df_pandas, tmpdir):
+    filepath = os.path.join(tmpdir, 'sample')
+    sample_df_pandas.to_feather(filepath)
+
+    content_type = "application/feather"
+    not_supported_error = "Reading from content type {} is not currently supported".format(content_type)
+    with pytest.raises(RuntimeError, match=not_supported_error):
+        ww.read_file(filepath=filepath, content_type=content_type)
+
+
+def test_read_file_uses_supplied_content_type(sample_df_pandas, tmpdir):
+    filepath = os.path.join(tmpdir, 'sample')
+    sample_df_pandas.to_csv(filepath, index=False)
+
+    df_from_csv = ww.read_file(filepath=filepath, content_type='csv')
     assert isinstance(df_from_csv.ww.schema, ww.table_schema.TableSchema)
 
     schema_df = sample_df_pandas.copy()
@@ -182,7 +201,25 @@ def test_read_csv_no_params(sample_df_pandas, tmpdir):
     pd.testing.assert_frame_equal(schema_df, df_from_csv)
 
 
-def test_read_csv_with_woodwork_params(sample_df_pandas, tmpdir):
+def test_read_file_no_params(sample_df_pandas, tmpdir):
+    filepath = os.path.join(tmpdir, 'sample.csv')
+    sample_df_pandas.to_csv(filepath, index=False)
+
+    df_from_csv = ww.read_file(filepath=filepath)
+    assert isinstance(df_from_csv.ww.schema, ww.table_schema.TableSchema)
+
+    schema_df = sample_df_pandas.copy()
+    # pandas does not read data into nullable types currently, so the types
+    # in df_from_csv will be different than the types inferred from sample_df_pandas
+    # which uses the nullable types
+    schema_df = schema_df.astype({'age': 'float64', 'is_registered': 'object'})
+    schema_df.ww.init()
+
+    assert df_from_csv.ww.schema == schema_df.ww.schema
+    pd.testing.assert_frame_equal(schema_df, df_from_csv)
+
+
+def test_read_file_with_woodwork_params(sample_df_pandas, tmpdir):
     filepath = os.path.join(tmpdir, 'sample.csv')
     sample_df_pandas.to_csv(filepath, index=False)
     logical_types = {
@@ -195,11 +232,11 @@ def test_read_csv_with_woodwork_params(sample_df_pandas, tmpdir):
         'age': ['tag1', 'tag2'],
         'is_registered': ['tag3', 'tag4']
     }
-    df_from_csv = ww.read_csv(filepath=filepath,
-                              index='id',
-                              time_index='signup_date',
-                              logical_types=logical_types,
-                              semantic_tags=semantic_tags)
+    df_from_csv = ww.read_file(filepath=filepath,
+                               index='id',
+                               time_index='signup_date',
+                               logical_types=logical_types,
+                               semantic_tags=semantic_tags)
     assert isinstance(df_from_csv.ww.schema, ww.table_schema.TableSchema)
 
     schema_df = sample_df_pandas.copy()
@@ -212,12 +249,12 @@ def test_read_csv_with_woodwork_params(sample_df_pandas, tmpdir):
     pd.testing.assert_frame_equal(schema_df, df_from_csv)
 
 
-def test_read_csv_with_pandas_params(sample_df_pandas, tmpdir):
+def test_read_file_with_pandas_params(sample_df_pandas, tmpdir):
     filepath = os.path.join(tmpdir, 'sample.csv')
     sample_df_pandas.to_csv(filepath, index=False)
     nrows = 2
 
-    df_from_csv = ww.read_csv(filepath=filepath, nrows=nrows, dtype={'age': 'Int64', 'is_registered': 'boolean'})
+    df_from_csv = ww.read_file(filepath=filepath, nrows=nrows, dtype={'age': 'Int64', 'is_registered': 'boolean'})
     assert isinstance(df_from_csv.ww.schema, ww.table_schema.TableSchema)
 
     schema_df = sample_df_pandas.copy()
@@ -229,14 +266,14 @@ def test_read_csv_with_pandas_params(sample_df_pandas, tmpdir):
 
 
 @patch("woodwork.table_accessor._validate_accessor_params")
-def test_read_csv_validation_control(mock_validate_accessor_params, sample_df_pandas, tmpdir):
+def test_read_file_validation_control(mock_validate_accessor_params, sample_df_pandas, tmpdir):
     filepath = os.path.join(tmpdir, 'sample.csv')
     sample_df_pandas.to_csv(filepath, index=False)
 
     assert not mock_validate_accessor_params.called
-    ww.read_csv(filepath=filepath, validate=False)
+    ww.read_file(filepath=filepath, validate=False)
     assert not mock_validate_accessor_params.called
-    ww.read_csv(filepath=filepath)
+    ww.read_file(filepath=filepath)
     assert mock_validate_accessor_params.called